split tabspage into tabs and tabspage
