--- src/components/TabsPage/TabsPage.tsx
@@ -1,39 +1,0 @@
-import { Link, useParams } from 'react-router-dom';
-import { Tab } from '../../types/Tab';
-
-type Props = {
-  tabs: Tab[];
-};
-
-export const TabsPage: React.FC<Props> = ({ tabs }) => {
-  const { tabId } = useParams();
-
-  const activeTab = tabs.find(tab => tab.id === tabId);
-
-  return (
-    <>
-      <h1 className="title">Tabs page</h1>
-
-      <div data-cy="TabsComponent">
-        <div className="tabs is-boxed">
-          <ul>
-            {tabs.map(({ id, title }) => (
-              <li
-                className={tabId === id ? 'is-active' : ''}
-                data-cy="Tab"
-                key={id}
-              >
-                <Link to={`/tabs/${id}`} data-cy="TabLink">
-                  {title}
-                </Link>
-              </li>
-            ))}
-          </ul>
-        </div>
-        <div className="block" data-cy="TabContent">
-          {activeTab ? activeTab.content : 'Please select a tab'}
-        </div>
-      </div>
-    </>
-  );
-};
